Fixing commit id check
