+++ vars/upperEnvsBuildAndDeploy.groovy
@@ -15,8 +15,6 @@ import com.ft.jenkins.slack.SlackUtils
 
 import static com.ft.jenkins.DeploymentUtilsConstants.HELM_CONFIG_FOLDER
 
-import groovy.json.JsonSlurper
-
 def call(GithubReleaseInfo releaseInfo, BuildConfig buildConfig) {
 
   DeploymentUtils deployUtil = new DeploymentUtils()
@@ -270,10 +268,7 @@ private String openCr(String approver, GithubReleaseInfo releaseInfo, Environmen
       data.environment = ChangeRequestEnvironment.Test
     }
     data.notifyChannel = environment.slackChannel
-    //this will be removed
-    def output = getGithubLatestCommit("master", "content-k8s-provisioner")
-    print "Latest commit hash of content-k8s-provisioner branch master is ${output}"
-    //this will be removed
+
     ChangeRequestsUtils crUtils = new ChangeRequestsUtils()
     return crUtils.open(data)
   }
@@ -291,22 +286,6 @@ final class JenkinsDeployInput implements Serializable {
     this.approver = approver
     this.selectedRegion = selectedRegion
   }
-}
-
-private String getGithubLatestCommit(String branch, String repoName) {
-  /*  fetch the release info*/
-  GString requestedUrl = "https://api.github.com/repos/Financial-Times/${repoName}/commits/${branch}"
-  try {
-    releaseResponse = httpRequest(acceptType: 'APPLICATION_JSON',
-                                  authentication: 'ft.github.credentials',
-                                  url: requestedUrl)
-  } catch (IllegalStateException e) {
-    echo"Release in GitHub could not be found at URL: ${requestedUrl}. Error: ${e.message}"
-    return null
-  }
-  def releaseInfoJson = new JsonSlurper().parseText(releaseResponse.content)
-  String latestCommit = releaseInfoJson.sha
-  return latestCommit
 }
 
 void sendSlackMessageForIntermediaryDeploy(GithubReleaseInfo releaseInfo, Environment targetEnv,
